+++ CHANGELOG.md
@@ -253,6 +253,8 @@ format.
 - Rewrite document-relative script URLs inside static export iframe `srcdoc`
   content so embedded page and asset references point at generated files.
   ([#3619](https://github.com/WordPress/wordpress-playground/pull/3619))
+- Treat microdata `itemid` attributes as crawlable static export page URLs.
+  ([#3619](https://github.com/WordPress/wordpress-playground/pull/3619))
 
 ### Documentation